==> codes/02_Twitter_textcleaning.ipynb <==
# 02. Twitter Text Cleaning

---

<h1>Table of Contents<span class="tocSkip"></span></h1>

- [1. Import Packages](#1.-Import-Packages)
- [2. Read in Twitter Data](#2.-Read-in-Twitter-Data)
- [3. Frame Cleaning](#3.-Frame-Cleaning)
- [4. Text Cleaning](#3.-Text-Cleaning)
 - [4A. Initial HTML Cleaning](#4A.-Initial-HTML-Cleaning)
 - [4B. Links Cleaning](#4B.-Links-Cleaning)
 - [4C. Additional HTML and Other Text Cleaning](#4C.-Additional-HTML-and-Other-Text-Cleaning)
- [5. Read to CSV](#5.-Read-to-CSV)

---

# 1.  Import Packages

import pandas as pd
import numpy as np
import requests
from nltk.stem import WordNetLemmatizer
import nltk
nltk.download('wordnet')
nltk.download('stopwords')
from nltk.tokenize import RegexpTokenizer
import regex as re
from nltk.stem.porter import PorterStemmer
from bs4 import BeautifulSoup  
from sklearn.feature_extraction import stop_words
import matplotlib.pyplot as plt
from nltk.corpus import stopwords
import warnings
warnings.filterwarnings('ignore')
%matplotlib inline

---

# 2. Read in Twitter Data

You can read in a csv file containing scraped twitter data that we will then run through to clean.  Please import the file path into string object that currently contains "./data/test.csv".

data = pd.read_csv('../datasets/scraped_tweets.csv')

data.head()

---

# 3. Frame Cleaning

data.drop(columns = 'Unnamed: 0', inplace = True)

data.head()

data.dtypes

data['date_posted'] = pd.to_datetime(data['date_posted'])

data.dtypes

---

# 4. Text Cleaning

## 4A. Initial HTML Cleaning

First off, we want to run a simiple Beautiful Soup object that begins to clear out some of the legacy HTML characters.

clean_text = [BeautifulSoup(data.loc[row,'tweet'],'lxml').text for row in list(range(data.shape[0]))]
data['tweet'] = clean_text

data.head()

## 4B. Links Cleaning

Next, we'll work to remove all links from the text.

# the below regex code replaces the links that being in http in our text
data['tweet'] = data.loc[:, 'tweet'].map(lambda row : re.sub(r'http\S+', '', row))
data['tweet'] = data.loc[:, 'tweet'].map(lambda row : re.sub(r'@\S+', '', row))

## 4C. Additional HTML  and Other Text Cleaning

We found that while our Beautiful Soup object was pretty helpful in providing an initial clean to the HTML characters, we discover in looking at our text items that there are still some characters left over.  Although some could be removed just with blanks, we will actually replace these with spaces so that original formatting is maintained.

# list of leftover HTML characters to clean out with blanks
rem_chars = ['\n', '\ufeff', '>', '**', '\'ve', '#', '…']

for char in rem_chars:
    data['tweet'] = data.loc[:, 'tweet'].map(lambda text : text.replace(char, ' '))

---

# 5. Read to CSV

data.to_csv('../datasets/clean_twitter.csv')

---
