==> codes/03_Pre-processing_and_EDA.ipynb <==
# 03. Pre-processing and EDA
---

The prurpose of this notebook is to prepare our web-scrapped data from the file [scraped_tweets.csv](#http://localhost:8888/edit/PROJECTS/project_5/project_5/datasets/scraped_tweets.csv) for future modeling.

---
# Table of Content #

- [1. Importing Libraries and Data](#1.-Importing-Libraries-and-Data)
- [2. Pre-processing](#2.-Pre-processing)
- [3. EDA](#3.-EDA)

---
## 1. Importing Libraries and Data ##

#Importing Libraries
import pandas as pd
import numpy as np

#Importing our data to a DataFrame
df = pd.read_csv('../datasets/clean_twitter.csv')

---
## 2. Pre-processing ##

#Getting data dimesions
df.shape

#Initial data check
df.head()

In order to determine what tweets are relevant to our analysis, we have conducted some code words filtering - see the details below.

#Creating individual masks to filter our DataFrame

mask_road = df['tweet'].str.contains('road', regex=False, case=False)
mask_exit = df['tweet'].str.contains('exit', regex=False, case=False)
mask_street = df['tweet'].str.contains('street', regex=False, case=False)
mask_highway = df['tweet'].str.contains('highway', regex=False, case=False)
mask_hwy = df['tweet'].str.contains('hwy', regex=False, case=False)
mask_ramp = df['tweet'].str.contains('ramp', regex=False, case=False)
mask_st = df['tweet'].str.contains('st', regex=False, case=False)
mask_ave = df['tweet'].str.contains('ave', regex=False, case=False)
mask_lane = df['tweet'].str.contains('lane', regex=False, case=False)
mask_ln = df['tweet'].str.contains('ln', regex=False, case=False)
mask_drive = df['tweet'].str.contains('drive', regex=False, case=False)
mask_clos = df['tweet'].str.contains('clos', regex=False, case=False)

#Creating a unified mask
mask = (mask_road |
                  mask_exit | 
                  mask_street | 
                  mask_highway | 
                  mask_hwy | 
                  mask_ramp | 
                  mask_st | 
                  mask_lane)& mask_clos

#Separating tweets dealing with actual road closures into a separate DataFrame
df_closures = df[mask]

#Separating tweets NOT dealing with actual road closures into a separate DataFrame
df_spam = df[~mask]

---
## 3. EDA ##

#Getting data dimensions for the closures DataFrame
df_closures.shape

#Checking for empty tweets
df_closures['tweet'].isna().sum()

#Getting data dimensions for the spam DataFrame
df_spam.shape

#Checking for empty tweets
df_spam['tweet'].isna().sum()

#Introducing a label for our future positive class
df_closures['mark'] = 0

#Introducing a label for our future negative class
df_spam['mark'] = 1

#Having a look at our spam DataFrame
df_spam

#Having a look at our road closures tweets DataFrame
df_closures.head()

#Dropping unnecessary and ininformative columns
df_closures = df_closures[['tweet','mark']]
df_spam = df_spam[['tweet','mark']]

#Resetting indices
df_closures.reset_index(inplace=True)
df_spam.reset_index(inplace=True)

#Getting an impression of a spam tweet
df_spam['tweet'][5]

#Getting an impression of a closures tweet
df_closures['tweet'][5]

In order to avoid inbalanced clasees with our future classification model we need to address this problem by constructing our modeling dataset in a way our positive class (field 'mark' = 0, tweets dealing with closed road parts) is not dissoluted in our negative class (field 'mark' = 1, tweets dealing with no road closures/partical closures - the idea is that for emergency planning it's important if a first-responce vehicle can use at least a part of the road). Hence, we will randomly pull tweets out of our negative 'spam' DataFrame - and a number of these tweets will be equal to the number of tweets in our positive DataFrame. 

#Randomly sampling our negative tweets into a new 'Negative' DataFrame
df_neg = df_spam.sample(df_closures.shape[0])

#Concatenating our positive and negative dataframes into the final one
df_final = pd.concat([df_closures, df_neg])

#Checking the final DataFrame dimensions
df_final.shape

#Saving our final DataFrame into a csv file for future modeling
df_final.to_csv('../datasets/modeling_df.csv')

---
