==> codes/06_Coordinates_Acquisition.ipynb <==
# 06. Coordinate Acquisition
---

The purpose of this notebook is to leverage the `geocoder` mapping library in order to acquire the latitude and longitude coordinates for the locations referenced in the collected and cleaned tweets.  This process is executed after we have extracted locations identified in each tweet posted by one of the NJ 511 twitter accounts in the previous notebook.

---
## Table of Contents
---

- [Import Resources](#Import-Resources)
- [DataFrame Series Setup](#DataFrame-Series-Setup)
- [Organize and Save](#Organize-and-Save)

---
### Import Resources
---

Here we import the necessary libraries to enable geocode acquisition.  Furthermore, we read-in the cleaned `closure_location` data.

import pandas as pd
import geocoder

Read-in and light cleaning of csv

locs_df = pd.read_csv("../datasets/closure_locations.csv")

# Drop unncessary column
locs_df.drop(columns = "Unnamed: 0", inplace = True)

# Fill blank cells with ''
locs_df.fillna('', inplace = True)

locs_df.head()

---
### DataFrame Series Setup
---

Setting up table for acquiring coordinates

Creating `state` variable to add to location

state = "NJ"

Iterating through each identifiable location and setting up a `location_lookup` column for each non `''` cell.

# creating state variable to add to location_lookup
state = 'NJ'

# creating blank location_lookup column to fill via below for loop
locs_df['location_lookup'] = ''

# loop to populate location_lookup column
for i in list(range(locs_df.shape[0])):
    if locs_df['town'].loc[i] != '':
        locs_df['location_lookup'].loc[i] = locs_df['town'].loc[i] + ', ' + state
        
    elif locs_df['exits'].loc[i] != '':
        locs_df['location_lookup'].loc[i] = locs_df['exits'].loc[i] + ', ' + state
        
    elif locs_df['streets'].loc[i] != '':
        locs_df['location_lookup'].loc[i] = locs_df['streets'].loc[i] + ', ' + state
        
    elif locs_df['highway'].loc[i] != '':
        locs_df['location_lookup'].loc[i] = locs_df['highway'].loc[i] + ', ' + state
        
    else:
        locs_df['location_lookup'].loc[i] = state

locs_df.head()

# test
geocoder.arcgis(locs_df['location_lookup'][0]).latlng

Acquiring coordinates - ***this will take about 30 minutes, given the amount of locations to iterate through***

locs_df['coordinates'] = [geocoder.arcgis(locs_df['location_lookup'][i]).latlng for i in list(range(locs_df.shape[0]))]

locs_df.head()

---
### Organize and Save
---

We will organize the results of the geocoder into a usable dataframe and save it down as a csv for use by other notebooks.

Splitting latitude and longitude into seperate Series in the DataFrame

# Create latitude column
locs_df['lat'] = [coords[0] for coords in locs_df['coordinates']]

# Create longitude column
locs_df['lon'] = [coords[1] for coords in locs_df['coordinates']]

# Drop coordinates column
locs_df.drop(columns = "coordinates", inplace = True)

# Take a look at results from above seperation
locs_df.head()

Saving the resulting DataFrame as a csv file for the next notebook to use

locs_df.to_csv('../datasets/geodata.csv')

---
