==> codes/07_Folium_Mapping.ipynb <==
# 07. Mapping
---

We will use this notebook to plot out the results of our efforts, in order to have an easy to interpret visual representation of all the data gathered.

---
## Table of Contents
---


 - [Import Resources](#Import-Resources)
 - [Setting Up Data for Mapping](#Setting-Up-Data-for-Mapping)
 - [Create and Save Map](#Create-and-Save-Map)

---
### Import Resources
---

Here we import the necessary libraries to enable mapping.  Furthermore, we read-in the cleaned `geodata` data.

import pandas as pd
import folium
from folium.plugins import MarkerCluster, FeatureGroupSubGroup

Read-in and light cleaning of csv

geo_data = pd.read_csv("../datasets/geodata.csv")

# Drop unnecessary columns
geo_data.drop(columns = "Unnamed: 0", inplace = True)

geo_data.head()

---
### Setting up Data for Mapping
---

# Create a list of usernames for dictionary reference
usernames = list(set(geo_data['username']))

user_dict = {}

for user in usernames:
    user_dict[user] = geo_data[geo_data['username'] == user].reset_index(drop = True)

Building a color dictionary to assign colors to respective twitter accounts that appear on html map.

## Color dictionary for usernames
color_dict = {
    'NJTRANSIT_PVL' : 'black',
    '511njace' : 'gray',
    '511njgsp' : 'silver',
    'DRPA_PAandNJ' : 'whitesmoke',
    'NJTRANSIT_NBUS' : 'rosybrown',
    '511nji78' : 'firebrick',
    'NJTRANSIT_MOBO' : 'red',
    'NJTRANSIT_NJCL' : 'darksalmon',
    'NJTRANSIT_ME' : 'sienna',
    '511nji676' : 'sandybrown',
    '511nji287' : 'bisque',
    'NJTRANSIT_MBPJ' : 'tan',
    '511nji280' : 'moccasin',
    '511nj24' : 'floralwhite',
    '511nji195' : 'gold',
    '511nj42' : 'darkkhaki',
    '511njtpk' : 'lightgoldenrodyellow',
    'NJTRANSIT' : 'olivedrab',
    'NJTRANSIT_SBUS' : 'chartreuse',
    'NJTRANSIT_RVL' : 'palegreen',
    'NJTurnpike' : 'darkgreen',
    'demembridge' : 'seagreen',
    'NJDOT_info' : 'mediumspringgreen',
    '511nji295' : 'lightseagreen',
    'NJTRANSIT_NEC' : 'paleturquoise',
    'NJTRANSIT_HBLR' : 'darkcyan',
    '511nj55' : 'deepskyblue',
    '511nj495' : 'aliceblue',
    'NJTRANSIT_RL' : 'navy',
    'NJTRANSIT_NLR' : 'blue',
    '511nji80' : 'plum',
    '511njbt' : 'm',
    '511nji76' : 'palevioletred'
}

## Add color column to each user dataframe with the color for that user
for user in user_dict:
    user_dict[user]['color'] = color_dict[user]

---
### Create and Save Map
---

## Adapted from Noah Chirstiansen's code

## Instantiate our map object
my_map = folium.Map(location = [40.717700, -74.043110], ## initial map focus
                    zoom_start = 10, ## zoom level
                    control_scale=True, 
                    prefer_canvas = True)

## Create the html for our popups
def popup_url(point):
    # Add username
    popup_html  = f'<h3>{point["username"]}</h3><br>'
    # Add tweet
    popup_html += f'<b>Description:</b> {point["tweet"]} <br>'
    # Add date posted
    popup_html += f'<b>Date/Time:</b> {point["date_posted"]} <br>'

    return popup_html

## Function to add a point to our map
def add_point(point):
    ## Parse the html we generate with the 'popup_url' function
    html_string = folium.Html(popup_url(point), script=True)
    
    ## Add a circle marker for the point
    folium.vector_layers.CircleMarker(location=[point['lat'], point['lon']],
                        color = point['color'], #set color using the color column
                        fill = True,             
                        fill_color = point['color'],
                        fill_opacity = 0.7,
                        radius=7,
                        weight=2,
                        popup = folium.Popup(html_string,max_width=500, parse_html=True), ## add our popup text to the point
                        control = True,
                        overlay = True).add_to(sub_group) ## add to the subgroup layer for that point

for user in user_dict:
    ## Instantiate MarkerCluster for each layer (this will cluster points together at higher zoom levels)
    marker_cluster = MarkerCluster(control = False)

    ## Add MarkerCluster to our map
    marker_cluster.add_to(my_map)

    ## Instantiate subgroup for each layer (this will group together all points in each dataframe)
    sub_group = FeatureGroupSubGroup(marker_cluster, name = user, control= True, show = True)

    ## Use apply to loop through our dataframe and add each point to our subgroup layer
    user_dict[user].apply(add_point, axis = 1)

    ## Add subgroup layer to our map
    sub_group.add_to(my_map)

## Add LayerControl to our map (this will allow us to select layers to display on our map)
folium.map.LayerControl(collapsed=False).add_to(my_map)

## Save map to html file
my_map.save('../maps/my_map.html')

---
